+++ src/Halcyon.Api/Models/Account/RegisterRequestValidator.cs
@@ -1,18 +1,18 @@
 ﻿using FluentValidation;
-using Halcyon.Api.Extensions;
 using Halcyon.Api.Services.Date;
+using Halcyon.Api.Validators;
 
 namespace Halcyon.Api.Models.Account
 {
     public class RegisterRequestValidator : AbstractValidator<RegisterRequest>
     {
-      public RegisterRequestValidator(IDateService dateService)
+      public RegisterRequestValidator(IDateTimeProvider dateTimeProvider)
         {
             RuleFor(x => x.EmailAddress).NotEmpty().EmailAddress().MaximumLength(255).WithName("Email Address");
             RuleFor(x => x.Password).NotEmpty().MinimumLength(8).MaximumLength(50);
             RuleFor(x => x.FirstName).NotEmpty().MaximumLength(50).WithName("First Name");
             RuleFor(x => x.LastName).NotEmpty().MaximumLength(50).WithName("Last Name");
-            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateService).WithName("Date Of Birth");
+            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateTimeProvider).WithName("Date Of Birth");
         }
     }
 }
