+++ src/Halcyon.Api/Models/Manage/UpdateProfileRequestValidator.cs
@@ -1,17 +1,17 @@
 ﻿using FluentValidation;
-using Halcyon.Api.Extensions;
 using Halcyon.Api.Services.Date;
+using Halcyon.Api.Validators;
 
 namespace Halcyon.Api.Models.Manage
 {
     public class UpdateProfileRequestValidator : AbstractValidator<UpdateProfileRequest>
     {
-        public UpdateProfileRequestValidator(IDateService dateService)
+        public UpdateProfileRequestValidator(IDateTimeProvider dateTimeProvider)
         {
             RuleFor(x => x.EmailAddress).NotEmpty().EmailAddress().MaximumLength(255).WithName("Email Address");
             RuleFor(x => x.FirstName).NotEmpty().MaximumLength(50).WithName("First Name");
             RuleFor(x => x.LastName).NotEmpty().MaximumLength(50).WithName("Last Name");
-            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateService).WithName("Date Of Birth");
+            RuleFor(x => x.DateOfBirth).NotEmpty().InThePast(dateTimeProvider).WithName("Date Of Birth");
         }   
     }
 }
