+++ src/Halcyon.Api/Models/Seed/SeedUser.cs
@@ -1,0 +1,17 @@
+﻿namespace Halcyon.Api.Models.Seed
+{
+    public class SeedUser
+    {
+        public string EmailAddress { get; set; }
+
+        public string Password { get; set; }
+
+        public string FirstName { get; set; }
+
+        public string LastName { get; set; }
+
+        public DateOnly DateOfBirth { get; set; }
+
+        public List<string> Roles { get; set; }
+    }
+}
