+++ src/Halcyon.Api/Program.cs
@@ -141,7 +141,7 @@ builder.Services.Configure<EmailSettings>(builder.Configuration.GetSection(Email
 builder.Services.Configure<JwtSettings>(builder.Configuration.GetSection(JwtSettings.SectionName));
 builder.Services.Configure<SeedSettings>(builder.Configuration.GetSection(SeedSettings.SectionName));
 
-builder.Services.AddSingleton<IDateService, DateService>();
+builder.Services.AddSingleton<IDateTimeProvider, DateTimeProvider>();
 builder.Services.AddSingleton<IHashService, HashService>();
 builder.Services.AddSingleton<IJwtService, JwtService>();
