+++ src/Halcyon.Api/Settings/SeedSettings.cs
@@ -1,24 +1,11 @@
-﻿namespace Halcyon.Api.Settings
+﻿using Halcyon.Api.Models.Seed;
+
+namespace Halcyon.Api.Settings
 {
     public class SeedSettings
     {
         public static string SectionName { get; } = "Seed";
 
         public List<SeedUser> Users { get; set; }
-    }
-
-    public class SeedUser
-    {
-        public string EmailAddress { get; set; }
-
-        public string Password { get; set; }
-
-        public string FirstName { get; set; }
-
-        public string LastName { get; set; }
-
-        public DateOnly DateOfBirth { get; set; }
-
-        public List<string> Roles { get; set; }
     }
 }
